Draft implementation of LoadGamePort in JpaRepositoryAdapter

+++ src/main/java/de/htwberlin/casino/blackjack/adapter/in/web/ChancesController.java
@@ -36,7 +36,7 @@ public class ChancesController {
             @ApiResponse(responseCode = "404", description = "Game not found", content = @Content)
     })
     @GetMapping("/chances/{gameId}")
-    public ResponseEntity<?> calculateChances(@PathVariable int gameId) {
+    public ResponseEntity<?> calculateChances(@PathVariable Long gameId) {
         var query = new CalculateChancesCommand(gameId);
         var result = calculateChancesUseCase.calculateChances(query);
