+++ src/main/java/de/htwberlin/casino/blackjack/application/domain/model/Game.java
@@ -5,15 +5,15 @@ import lombok.Getter;
 @Getter
 public class Game {
 
-    private final int gameId;
-    private final int userId;
+    private final Long gameId;
+    private final Long userId;
     private CardDeck cardDeck;
     private PlayerHand playerHand;
     private DealerHand dealerHand;
     private GameState gameState;
     private final double bet;
 
-    public Game(int gameId, int userId, CardDeck cardDeck,
+    public Game(Long gameId, Long userId, CardDeck cardDeck,
                 PlayerHand playerHand, DealerHand dealerHand, GameState gameState, double bet) {
         this.gameId = gameId;
         this.userId = userId;
@@ -24,7 +24,7 @@ public class Game {
         this.bet = bet;
     }
 
-    public Game(int gameId, int userId, double bet) {
+    public Game(Long gameId, Long userId, double bet) {
         this.gameId = gameId;
         this.userId = userId;
         this.bet = bet;
